Handle config import errors
