+++ src/ume/bootstrap/__init__.py
@@ -1,0 +1,13 @@
+"""Helpers for bootstrapping the UME package."""
+
+from .config import load_config
+from .neo4j import load_neo4j
+from .vector import load_vector_modules
+from .embedding import load_embedding
+
+__all__ = [
+    "load_config",
+    "load_neo4j",
+    "load_vector_modules",
+    "load_embedding",
+]
